backend/requirements.txt
# RAG Chatbot - Semantic Retrieval Pipeline
# Python 3.10+

# Framework & LLM
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
openai>=1.0.0

# Document loading
pypdf>=4.0.0
python-docx>=1.0.0
docling>=2.0.0

# Vector store: Qdrant 
qdrant-client>=1.8.0

# Embeddings & re-ranking
sentence-transformers>=2.2.0
transformers>=4.40.0
accelerate>=0.28.0
torch>=2.2.0

# Token counting
tiktoken>=0.5.0

# API & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
slowapi>=0.1.9

# HTTP client (drug price tool — direct pharmacy scraping)
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Persistence & auth
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
